feat(title-with-refetch): add component to display title with refetch button
This change introduces a new component called `TitleWithRefetch` that
displays a title with a refetch button. The component takes in a `refetch`
function, an `isLoading` flag, and the title content as children. The
refetch button is disabled while the component is loading.

The component is designed to be used in places where the user needs to
be able to refresh the data being displayed, such as in a list or
detail view.