feat(ci): never cancel workflows on non-PRs, add support for running workflows on tags and merge groups (#167)
The changes made in this commit add support for running the lint and test workflows on both the main branch and on tags that start with 'v'. Additionally, the concurrency settings have been updated to only cancel in-progress runs for pull_request events, preventing the cancellation of workflows against the main branch or tags.