feat(library): initial library implementation (#118)
* feat(library): initial library implementation

This commit adds the ability to create, scan, and fetch libraries in the server-side application. The key changes are:

- Adds a new `library` table in the database schema to store library information.
- Implements tRPC procedures for creating a new library, scanning a library, and fetching all libraries.
- Adds a library scanning machine using the XState library to handle the asynchronous scanning process.
- Integrates the Audible API to search for and match audio files to Audible products.
- Adds a new `LibraryListScreen` component in the client application to manage libraries.

The main goal of these changes is to provide a way for users to manage their audio libraries, including the ability to scan and identify the audio files in their libraries.

* feat(db): add new database tables for library management

This commit introduces a set of new database tables to support the library management
functionality of the application. The changes include:

- Creation of the `author`, `series`, `book`, `bookAuthor`, `bookSeries`, `bookContributor`,
  `audiobookChapter`, `audiobookFile`, and `ebookFile` tables, along with their respective
  triggers to update the `updatedAt` column on changes.
- Addition of foreign key constraints and indexes to ensure data integrity and performance.
- Implementation of strict schema enforcement using the `STRICT` modifier.

These changes lay the foundation for managing authors, book series, books, and their
associated metadata, as well as audiobook and ebook files, within the application's
library management system.

* finish up initial implementation and solar-core integration